+++ crates/ghost-linux/src/system/mod.rs
@@ -3,4 +3,4 @@ pub mod dpi;
 pub mod window;
 
 pub use clipboard::{get_clipboard, set_clipboard};
-pub use window::{cursor_pos, foreground_window, foreground_window_rect, window_rect};
+pub use window::{window_pid, cursor_pos, foreground_window, foreground_window_rect, window_rect};